@@ -13,71 +13,72 @@
     "cover": "nyc mocha"
   },
   "devDependencies": {
-    "@types/chai": "^3.5.2",
-    "@types/chai-enzyme": "^0.6.1",
-    "@types/classnames": "^2.2.0",
-    "@types/enzyme": "^2.8.0",
-    "@types/glob": "^5.0.30",
-    "@types/history": "^4.5.1",
-    "@types/mocha": "^2.2.41",
-    "@types/node": "^7.0.22",
-    "@types/node-fetch": "^1.6.7",
-    "@types/prop-types": "^15.5.1",
-    "@types/query-string": "^4.3.1",
-    "@types/react": "^15.0.25",
-    "@types/react-dom": "^15.5.0",
-    "@types/react-redux": "^4.4.40",
-    "@types/react-router": "^4.0.9",
-    "@types/react-router-dom": "^4.0.4",
-    "@types/react-router-redux": "^5.0.1",
-    "@types/react-tap-event-plugin": "^0.0.30",
-    "@types/recompose": "^0.22.0",
-    "@types/redux-immutable-state-invariant": "^2.0.0",
-    "@types/redux-logger": "^3.0.0",
-    "@types/sinon": "^2.2.2",
-    "chai": "^4.0.0",
-    "chai-enzyme": "^0.7.1",
-    "cross-env": "^5.0.0",
+    "@types/chai": "^4.2.16",
+    "@types/chai-enzyme": "^0.6.8",
+    "@types/classnames": "^2.2.11",
+    "@types/enzyme": "^3.10.8",
+    "@types/glob": "^7.1.3",
+    "@types/history": "^4.7.8",
+    "@types/mocha": "^8.2.2",
+    "@types/node": "^14.14.41",
+    "@types/node-fetch": "^2.5.10",
+    "@types/prop-types": "^15.7.3",
+    "@types/query-string": "^6.3.0",
+    "@types/react": "^17.0.3",
+    "@types/react-dom": "^17.0.3",
+    "@types/react-redux": "^7.1.16",
+    "@types/react-router": "^5.1.13",
+    "@types/react-router-dom": "^5.1.7",
+    "@types/react-router-redux": "^5.0.18",
+    "@types/recompose": "^0.30.7",
+    "@types/redux-immutable-state-invariant": "^2.1.1",
+    "@types/redux-logger": "^3.0.8",
+    "@types/sinon": "^10.0.0",
+    "chai": "^4.3.4",
+    "chai-enzyme": "^0.8.0",
+    "cheerio": "^1.0.0-rc.6",
+    "cross-env": "^7.0.3",
     "cue-parser-plus": "^0.0.5",
-    "enzyme": "^2.8.2",
-    "fuse-box": "2.1.0",
-    "mocha": "^3.4.2",
-    "nyc": "^10.3.2",
-    "react-addons-test-utils": "^15.5.1",
-    "sinon": "^2.3.2",
-    "ts-node": "^3.0.4",
-    "tslib": "^1.7.1",
-    "typescript": "^2.3.3",
-    "uglify-js": "^3.0.13"
+    "enzyme": "^3.11.0",
+    "fuse-box": "^4.0.1-next.4",
+    "mocha": "^8.3.2",
+    "nyc": "^15.1.0",
+    "sinon": "^10.0.0",
+    "ts-node": "^9.1.1",
+    "tslib": "^2.2.0",
+    "typescript": "^4.2.4",
+    "uglify-js": "^3.13.4"
   },
   "dependencies": {
-    "classnames": "^2.2.5",
-    "fetch-jsonp": "^1.0.7",
-    "history": "^4.6.1",
-    "material-ui": "1.0.0-alpha.16",
-    "material-ui-icons": "1.0.0-alpha.3",
-    "preact": "^8.1.0",
-    "preact-compat": "^3.16.0",
-    "preact-tap-event-plugin": "^0.1.2",
-    "react": "^15.5.4",
-    "react-dom": "^15.5.4",
-    "react-jss": "^6.1.1",
-    "react-redux": "^5.0.5",
-    "react-router": "^4.1.1",
-    "react-router-dom": "^4.1.1",
+    "@babel/polyfill": "^7.12.1",
+    "@babel/runtime": "^7.13.10",
+    "@material-ui/core": "^4.11.3",
+    "@material-ui/icons": "^4.11.2",
+    "classnames": "^2.3.1",
+    "fetch-jsonp": "^1.1.3",
+    "history": "^5.0.0",
+    "jss-preset-default": "^10.6.0",
+    "preact": "^10.5.13",
+    "preact-compat": "^3.19.0",
+    "preact-context": "^1.1.4",
+    "react": "^17.0.2",
+    "react-addons-test-utils": "^15.6.2",
+    "react-dom": "^17.0.2",
+    "react-jss": "^10.6.0",
+    "react-redux": "^7.2.3",
+    "react-router": "^5.2.0",
+    "react-router-dom": "^5.2.0",
     "react-router-redux": "5.0.0-alpha.6",
-    "react-tap-event-plugin": "^2.0.1",
-    "recompose": "0.23.4",
-    "redux": "^3.6.0",
-    "redux-devtools-extension": "^2.13.2",
-    "redux-immutable-state-invariant": "^2.0.0",
+    "redux": "^4.0.5",
+    "redux-devtools-extension": "^2.13.9",
+    "redux-immutable-state-invariant": "^2.1.0",
     "redux-logger": "^3.0.6",
-    "redux-thunk": "^2.2.0",
-    "reselect": "^3.0.1",
-    "smoothscroll-polyfill": "^0.3.5",
-    "typescript-fsa": "^2.3.0",
-    "typescript-fsa-reducers": "^0.4.3",
-    "typescript-fsa-redux-thunk": "^1.0.0"
+    "redux-thunk": "^2.3.0",
+    "reselect": "^4.0.0",
+    "smoothscroll-polyfill": "^0.4.4",
+    "typescript-fsa": "^3.0.0",
+    "typescript-fsa-reducers": "^1.2.2",
+    "typescript-fsa-redux-thunk": "^2.10.1"
   },
   "nyc": {
     "include": [
