@@ -1,44 +1,48 @@
-import actionCreatorFactory from 'typescript-fsa';
-import { bindThunkAction } from 'typescript-fsa-redux-thunk';
-import { State } from "../types/index";
+import actionCreatorFactory from "typescript-fsa";
+import { asyncFactory } from "typescript-fsa-redux-thunk";
+import { State } from "../types";
 import { MusicControl } from "../util/music";
 
-const create = actionCreatorFactory('music');
+const create = actionCreatorFactory("music");
+const createAsync = asyncFactory<State>(create);
 
-export const setControl = create<{ control: () => MusicControl }>('SET_CONTROL');
+export const setControl = create<{ control: () => MusicControl }>(
+  "SET_CONTROL"
+);
 
-export const playAsync = create.async<State, any, any, any>('PLAY');
-export const play = bindThunkAction(playAsync, async (_, __, getState) => {
-    const { control } = getState().music;
-    await control().play();
+export const play = createAsync("PLAY", async (_, __, getState) => {
+  const { control } = getState().music;
+  await control().play();
 });
 
-export const pauseAsync = create.async<State, any, any, any>('PAUSE');
-export const pause = bindThunkAction(pauseAsync, async (_, __, getState) => {
-    const { control } = getState().music;
-    await control().pause();
+export const pause = createAsync("PAUSE", async (_, __, getState) => {
+  const { control } = getState().music;
+  await control().pause();
 });
 
-export const stopAsync = create.async<State, any, any, any>('STOP');
-export const stop = bindThunkAction(stopAsync, async (_, __, getState) => {
-    const { control } = getState().music;
-    await control().stop();
+export const stop = createAsync("STOP", async (_, __, getState) => {
+  const { control } = getState().music;
+  await control().stop();
 });
 
-export const setSrcAsync = create.async<State, { src: string }, boolean, any>('SET_SRC');
-export const setSrc = bindThunkAction(setSrcAsync, async ({ src }, _, getState) => {
+export const setSrc = createAsync<{ src: string }, boolean, State>(
+  "SET_SRC",
+  async ({ src }, _, getState) => {
     const { control } = getState().music;
-    return control().setSourceUrl(src)
-});
+    return control().setSourceUrl(src);
+  }
+);
 
-export const setTimeAsync = create.async<State, { time: number }, any, any>('SET_TIME');
-export const setTime = bindThunkAction(setTimeAsync, async ({ time }, _, getState) => {
+export const setTime = createAsync<{ time: number }, any, State>(
+  "SET_TIME",
+  async ({ time }, _, getState) => {
     const { control } = getState().music;
     return control().setCurrentTime(time);
-});
+  }
+);
 
-export const setSeeking = create<{ seeking: boolean }>('SET_SEEKING');
-export const setPlaying = create<{ playing: boolean }>('SET_PLAYING');
-export const setWaiting = create<{ waiting: boolean }>('SET_WAITING');
-export const timeUpdate = create<{ currentTime: number }>('TIME_UPDATE');
-export const loadedMetadata = create<{ duration: number }>('LOADED_METADATA');
+export const setSeeking = create<{ seeking: boolean }>("SET_SEEKING");
+export const setPlaying = create<{ playing: boolean }>("SET_PLAYING");
+export const setWaiting = create<{ waiting: boolean }>("SET_WAITING");
+export const timeUpdate = create<{ currentTime: number }>("TIME_UPDATE");
+export const loadedMetadata = create<{ duration: number }>("LOADED_METADATA");
