@@ -1,6 +1,6 @@
-import * as musicActions from '../../../actions/music';
+import * as musicActions from "../../../actions/music";
 
 export const Controller = {
-    stop: musicActions.stop,
-    loadedMetadata: musicActions.loadedMetadata
+  stop: musicActions.stop,
+  loadedMetadata: musicActions.loadedMetadata,
 };
