@@ -1,16 +1,19 @@
-type Props = {
-    width: string;
-}
+// type Props = {
+//   width: string;
+// };
 
-export default {
-    title: {
-        'font-size': (props: Props) => {
-            if (props.width === 'xs') {
-                return '7.95px';
-            } else {
-                return '12px'
-            }
-        },
-        fontWeight: 'bold'
-    }
-};
+export default () => ({
+  title: {
+    color: "#fff",
+  },
+  mixList: {
+    // position: 'absolute',
+    // justifyContent: "flex-end",
+    // top: "45px",
+    // right: "0px",
+    // zIndex: 2,
+  },
+  active: {
+    // fontWeight: "bold",
+  },
+});
