@@ -1,10 +1,10 @@
-export const Stylesheet = {
-    appBar: {
-        position: 'relative',
-        height: '45px'
-    },
-    gridContainer: {
-        alignItems: 'center',
-        // justifyContent: 'space-between'
-    }
-};
+export const Stylesheet = () => ({
+  appBar: {
+    // position: "relative",
+    // height: "45px",
+  },
+  gridContainer: {
+    alignItems: "center",
+    justifyContent: "space-between",
+  },
+});
