@@ -1,6 +1,6 @@
-import * as musicActions from '../../../../actions/music';
+import * as musicActions from "../../../../actions/music";
 
 export const Controller = {
-    play: musicActions.play,
-    pause: musicActions.pause
+  play: musicActions.play,
+  pause: musicActions.pause,
 };
