@@ -1,11 +1,12 @@
-export const styles = {
+export const styles = () =>
+  ({
     track: {
-        position: 'absolute',
-        borderLeft: '1px dotted rgba(255,255,255,0.4)',
-        height: '100%',
-        color: '#fff'
+      position: "absolute",
+      borderLeft: "1px dotted rgba(255,255,255,0.4)",
+      height: "100%",
+      color: "#fff",
     },
     number: {
-        fontSize: '9px'
-    }
-};
+      fontSize: "9px",
+    },
+  } as const);
