@@ -1,5 +1,5 @@
-import * as musicActions from '../../../../actions/music';
+import * as musicActions from "../../../../actions/music";
 
 export const Controller = {
-    setTime: musicActions.setTime
+  setTime: musicActions.setTime,
 };
