@@ -1,46 +1,57 @@
-import * as React from 'react';
-import * as classNames from 'classnames';
-import { connect } from 'react-redux';
-import { injectCSS } from '../../../../util/jss';
-
-import Paper from 'material-ui/Paper';
-import { List, ListItem } from 'material-ui/List';
-
-import { Model } from './Model';
-import { Controller } from './Controller';
-import { ViewModel } from './ViewModel';
-import styles from './styles';
-import { zeroPad, secondsToTime2 } from "../../../../util/player";
+import { List, ListItem, makeStyles, Paper } from "@material-ui/core";
+import classNames from "classnames";
+import React from "react";
+import { useDispatch, useSelector } from "react-redux";
+import { bindActionCreators } from "redux";
+import { getCurrentTrack, getTracks } from "../../../../selectors/archive";
+import { State, Track } from "../../../../types";
+import { secondsToTime2, zeroPad } from "../../../../util/player";
 import ScrollToItem from "../../../util/ScrollToItem";
+import { Controller } from "./Controller";
+import { Props } from "./Model";
+import styles from "./styles";
 
-const C = connect(Model, Controller, ViewModel);
+const useStyles = makeStyles(styles);
 
-const View = C(({ classes, actions, track, tracks }) => tracks.length > 0 && (
-    <Paper className={classes.tracklist}>
+const View: React.FunctionComponent<Props> = (props) => {
+  const track = useSelector<State, Track>((state) =>
+    getCurrentTrack(state, props.mixId)
+  );
+  const tracks = useSelector<State, Track[]>((state) =>
+    getTracks(state, props.mixId)
+  );
+  const actions = bindActionCreators(Controller, useDispatch());
+  const classes = useStyles();
+  return (
+    (tracks.length > 0 && (
+      <Paper className={classes.tracklist}>
         <ScrollToItem
-            itemSelector={classes.track}
-            setHeight={() => `${window.innerHeight - (182)}px`}
+          itemSelector={classes.track}
+          setHeight={() => `${window.innerHeight - 182}px`}
         >
-            <List>
-                {track && tracks.map((t, index) => (
-                    <ListItem
-                        key={`track-${index}-${t.title}`}
-                        button
-                        className={classNames({
-                            [classes.track]: t.number === track.number
-                        })}
-                        onClick={() => actions.setTime({ time: t.time })}
-                    >
-                        <span>[{secondsToTime2(t.time)}]</span>
-                        &nbsp;
-                        <span>{zeroPad(t.number)}.</span>
-                        &nbsp;
-                        <span>{t.title}</span>
-                    </ListItem>
-                ))}
-            </List>
+          <List>
+            {track &&
+              tracks.map((t, index) => (
+                <ListItem
+                  key={`track-${index}-${t.title}`}
+                  button
+                  className={classNames({
+                    [classes.track]: t.number === track.number,
+                  })}
+                  onClick={() => actions.setTime({ time: t.time })}
+                >
+                  <span>[{secondsToTime2(t.time)}]</span>
+                  &nbsp;
+                  <span>{zeroPad(t.number)}.</span>
+                  &nbsp;
+                  <span>{t.title}</span>
+                </ListItem>
+              ))}
+          </List>
         </ScrollToItem>
-    </Paper>
-) || <div />);
+      </Paper>
+    )) || <div />
+  );
+};
 
-export default injectCSS(styles)(View);
+export default View;
