@@ -1,17 +1,16 @@
-import * as React from 'react';
-import Header from './Header';
-import MixList from './MixList'
-import Player from './Player';
-import { Route, RouteComponentProps, withRouter } from "react-router-dom";
+import React from "react";
+import Header from "./Header";
+import Player from "./Player";
+import { Route, useRouteMatch } from "react-router-dom";
 
-export type RouteProps = RouteComponentProps<{ mixId: string }>;
-
-const Index: React.SFC<RouteProps> = ({ match }) => (
+const Index: React.FunctionComponent = () => {
+  const match = useRouteMatch<{ mixId: string }>();
+  return (
     <div>
-        <Header mixId={match.params.mixId} />
-        <MixList mixId={match.params.mixId} />
-        <Route exact path={`/:mixId`} component={Player} />
+      <Header mixId={match.params.mixId} />
+      <Route exact path={`/:mixId`} component={Player} />
     </div>
-);
+  );
+};
 
-export default withRouter(Index);
+export default Index;
