@@ -1,23 +1,19 @@
-import * as React from 'react';
+import * as React from "react";
 
 type Props = {
-    src: string;
-    width?: number;
-    height?: number;
-    title?: string;
-    alt?: string;
-}
+  src: string;
+  width?: number;
+  height?: number;
+  title?: string;
+  alt?: string;
+};
 
-type Type = React.SFC<Props>;
+type Type = React.FunctionComponent<Props>;
 
-export const Icon: Type = props => {
-    const {
-        width = 20,
-        height = 20,
-        ...rest
-    } = props;
+export const Icon: Type = (props) => {
+  const { width = 20, height = 20, ...rest } = props;
 
-    return <img width={width} height={height} {...rest} />;
+  return <img width={width} height={height} {...rest} />;
 };
 
 export default Icon;
