@@ -1,16 +1,14 @@
-import { combineReducers, Reducer } from 'redux';
-import { routerReducer } from 'react-router-redux';
-import { State } from '../types';
-import { archive } from './archive';
-import { ui } from './ui';
-import { music } from './music';
+import { combineReducers, Reducer } from "redux";
+import { State } from "../types";
+import { archive } from "./archive";
+import { ui } from "./ui";
+import { music } from "./music";
 
 export const createReducer = (reducers?: { [idx: string]: Reducer<State> }) => {
-    return combineReducers({
-        archive,
-        ui,
-        music,
-        router: routerReducer,
-        ...reducers
-    }) as Reducer<State>;
+  return combineReducers({
+    archive,
+    ui,
+    music,
+    ...reducers,
+  }) as Reducer<State>;
 };
