@@ -1,19 +1,19 @@
-import * as fetchP from 'fetch-jsonp';
+import fetchP from "fetch-jsonp";
 
 export type TimeoutInit = RequestInit & { timeout?: number } & fetchP.Options;
 
 export interface FetchTimeout {
-    (url: string, options: TimeoutInit, f?: typeof fetch): Promise<Response>;
+  (url: string, options: TimeoutInit, f?: typeof fetch): Promise<Response>;
 }
 
 export const fetchT: FetchTimeout = (url, options, f = fetch) => {
-    const { timeout, ...opts } = options;
-    return Promise.race([
-        f(url, opts),
-        new Promise((_, reject) => {
-            setTimeout(() => {
-                reject(new Error('request timeout'));
-            }, timeout || 5000);
-        })
-    ]);
+  const { timeout, ...opts } = options;
+  return Promise.race([
+    f(url, opts),
+    new Promise<Response>((_, reject) => {
+      setTimeout(() => {
+        reject(new Error("request timeout"));
+      }, timeout || 5000);
+    }),
+  ]);
 };
